requirements.txt
python-telegram-bot==21.6
cryptography==43.0.3
python-dotenv==1.0.1
aiohttp==3.10.10
google-api-python-client==2.150.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
